@@ -44,7 +44,8 @@
     "@types/react": "18.2.8",
     "@types/react-dom": "18.2.4",
     "@types/react-signature-canvas": "^1.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/eslint-plugin": "^6.9.1",
+    "@typescript-eslint/parser": "^6.9.1",
     "@uploadthing/react": "^5.5.0",
     "autoprefixer": "10.4.14",
     "class-variance-authority": "^0.6.1",
@@ -54,11 +55,11 @@
     "dotenv": "^16.1.4",
     "drizzle-orm": "^0.27.0",
     "eslint": "^7.23.0",
-    "eslint-config-next": "^13.5.4",
     "framer-motion": "^10.16.4",
+    "geist": "^1.0.0",
     "lucide-react": "^0.233.0",
     "luxon": "^3.4.2",
-    "next": "^13.5.2",
+    "next": "^14.0.1",
     "next-themes": "^0.2.1",
     "postcss": "^8.4.31",
     "qrcode": "^1.5.3",
@@ -85,6 +86,7 @@
     "@types/eslint": "^8.44.6",
     "@types/react-html5-camera-photo": "^1.5.1",
     "drizzle-kit": "^0.19.3",
+    "eslint-config-next": "^14.0.1",
     "ngrok": "^5.0.0-beta.2"
   }
 }
