@@ -1,8 +1,64 @@
+import { Bars3Icon } from "@heroicons/react/24/outline";
+import Image from "next/image";
 import Link from "next/link";
+
+import ClientHeroSection from "./clientHero";
+
+const navigation = [
+  { name: "Home", href: "#", disabled: false },
+  { name: "About", href: "#about", disabled: false },
+  { name: "Shop (Coming Soon)", href: "#", disabled: true },
+  { name: "Contact Us", href: "#contact", disabled: false },
+];
 
 const Hero = () => {
   return (
     <>
+      <header className="absolute inset-x-0 top-0 z-50">
+        <nav
+          className="flex items-center justify-between p-6 lg:px-8"
+          aria-label="Global"
+        >
+          <div className="flex lg:flex-1">
+            <a href="#" className="-m-1.5 p-1.5">
+              <span className="sr-only">Your Company</span>
+              <Image
+                className="h-8 w-auto"
+                src="https://corruptedstrengthgym.com/cdn/shop/files/CS_Text-1_300x300.png?v=1621880690"
+                // src="/croppedMisfitsLogo.png"
+                alt=""
+                height={128}
+                width={128}
+              />
+            </a>
+          </div>
+          <div className="flex lg:hidden">
+            <ClientHeroSection />
+          </div>
+          <div className="hidden lg:flex lg:gap-x-12">
+            {navigation.map((item) => (
+              <Link
+                key={item.name}
+                href={item.href}
+                className={`text-sm font-semibold leading-6 text-[#EFE1B2] ${
+                  item.disabled ? "opacity-50 cursor-not-allowed" : null
+                }`}
+                aria-disabled={item.disabled}
+              >
+                {item.name}
+              </Link>
+            ))}
+          </div>
+          <div className="hidden lg:flex lg:flex-1 lg:justify-end">
+            <a
+              href="/sign-in"
+              className="text-sm font-semibold leading-6 text-gray-900"
+            >
+              Log in <span aria-hidden="true">&rarr;</span>
+            </a>
+          </div>
+        </nav>
+      </header>
       <div className="relative isolate px-6 pt-14 lg:px-8">
         <div
           className="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80"
@@ -19,22 +75,24 @@ const Hero = () => {
         <div className="mx-auto max-w-2xl py-32 sm:py-48 lg:py-56">
           <div className="text-center">
             <h1 className="text-4xl font-bold tracking-tight text-[#EFE1B2] sm:text-6xl">
-              Welcome Misfits!
+              HOME OF THE MISFITS
             </h1>
-            <p className="mt-6 text-lg leading-8 text-[#EFE1B2]">
-              Welcome to Corrupted Strength&apos;s Misfits Managed! A gym
-              management suite made for the members, by the members!
+            <p className="mt-6 text-lg leading-8 text-[#EFE1B2]/80">
+              {/* Welcome to Corrupted Strength&apos;s Misfits Managed! A gym
+              management suite made for the members, by the members! */}
+              Welcome to Corrupted Strength, your home for powerlifting in the
+              heart of Houston, Texas.
             </p>
             <div className="mt-10 flex items-center justify-center gap-x-6">
               <Link
                 href="/sign-up"
-                className="rounded-md bg-[#EA0607] px-3.5 py-2.5 text-sm font-semibold text-gray-900 shadow-sm hover:bg-[#EA0607]/90 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
+                className="rounded-md bg-[#EA0607] px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-[#EA0607]/90 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
               >
                 Sign Up
               </Link>
               <Link
                 href="/sign-in"
-                className="text-sm font-semibold leading-6 text-[#EFE1B2]"
+                className="text-sm font-semibold leading-6 text-white"
               >
                 Sign In <span aria-hidden="true">→</span>
               </Link>
